+++ site/src/Pages/AcessPage/Components/ComponentName/styleHeaders.css
@@ -1,0 +1,202 @@
+/*Estilos do header*/
+    /*Decide qual header usar*/
+    @media screen and (max-width: 1000px){
+        .headerMenor{
+            display: block;
+        }
+        .headerInicial{
+            display: none;
+        }
+    }
+    @media screen and (min-width: 1060px){
+        .headerMenor{
+            display: none;
+        }
+        .headerInicial{
+            display: block;
+        }
+    }
+
+    /*Colapsa a barra em algumas distâncias*/
+    @media screen and (max-width: 1100px){
+        .headerInicial nav ul li:nth-child(1){
+            display: none;
+        }
+    }
+
+    /*Configura o header da pagina inicial*/
+    header{
+        width: calc(100vw - 15px);
+        height: var(--alutraHeader);
+
+        position: fixed;
+        top: 0;
+        left: 0;
+
+        background: var(--preto); 
+
+        color: white;
+        font-weight: var(--fonteNormal);
+        font-size: 1.3rem;
+        z-index: 100;
+    }
+
+    /*Arruma a altura da navbar*/
+    nav{
+        height: inherit;
+        display: flex;
+        flex-direction: row;
+        justify-content: space-between;
+        align-items: center;
+        width: calc(100% - 8 * var(--margem));
+        margin: 0 calc(4*var(--margem));
+    }
+
+    /*Arruma o posicionamento do logo*/
+    .logo{
+        height: var(--alutraHeader);
+        overflow: visible;
+    }
+
+    /*Display em flexbox para os botões da página inicial*/
+    header nav ul{
+        list-style: none;
+        height: inherit;
+        display: flex;
+        flex-direction: row;
+        flex-wrap: wrap;
+        justify-content: right;
+        align-items: center;
+        float: right;
+        gap: calc(2*var(--margem));
+    }
+    
+    /*Estiliza todos os botões com exceção do botão de acesso*/
+    header nav ul li:not(header ul li:last-of-type){
+        vertical-align: center;
+        transition: 0.3s;
+        text-decoration: underline rgba(255,255,255,0);
+        text-underline-offset: 6px;
+    }
+
+    header nav ul li:not(header ul li:last-of-type):hover{
+        text-decoration: underline rgba(255,255,255,1);
+        font-weight: var(--fonteNegrito);
+        cursor: pointer;
+    }
+
+    /*Estiliza o botão de acesso*/
+    header nav ul li button{
+        height: 45px;
+        width: 120px;
+        
+        color: white;
+        font-weight: var(--fonteNormal);
+        font-family: Montserrat !important;
+        font-size: 1.3rem !important;
+
+        background-color: rgba(0,0,0,0);
+        border: white solid 1px;
+        border-radius: 5px;
+
+        transition: 0.5s;
+    }
+
+    header nav ul li button:hover{
+        font-weight: var(--fonteNegrito);
+        background-color: white;
+        color: black;
+        cursor: pointer;
+    }
+
+    /*Estiliza o menu lateral da página*/
+    .menuLateral{
+        width: 120px;
+    }
+
+    /*Posiciona os itens correspondentemente*/
+    .menuLateral ul{
+        list-style: none;
+        display: flex;
+        flex-direction: column;
+        justify-content: left;
+        align-items: flex-start;
+        gap: var(--margem);
+        float: left;
+    }
+
+    /*Hover do menu*/
+    .menuLateral .iconeMenu:hover{
+        cursor: pointer;
+    }
+
+    .menuLateral .iconeMenu:hover li:nth-child(2){
+        width: 20px;
+    }
+
+    /*cria as 3 barrinhas*/
+    .menuLateral .iconeMenu li{
+        width: 30px;
+        height: 4px;
+        background-color: white;
+        border-radius: 2px;
+        transition: 0.5s;
+    }
+
+    /*Transforma o menu lateral em um botão*/
+    .menuLateral .iconeMenu li button{
+        background-color: rgba(0,0,0,0);
+        width: 34px;
+        height: 36px;
+        transform: translate(-2px, -30px);
+        border: none;
+    }
+
+    /*Estiliza o menu escondido*/
+    .hiddenMenu{
+        position: fixed;
+        top: 80px;
+        left: 0;
+        width: calc(100vw - 15px);
+        height: auto;
+        box-sizing: border-box;
+        background-color: var(--preto);
+        overflow: hidden;
+        transition: 1s;
+        padding: 0;
+        opacity: 0;
+    }
+
+    .hiddenMenu ul{
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        flex-wrap: nowrap;
+        gap: 10px;
+        width: 100%;
+        list-style: none;
+    }
+
+    /*Underlines nas opções*/
+    .hiddenMenu ul li:hover{
+        text-decoration: underline rgba(255,255,255,1);
+        font-weight: var(--fonteNegrito);
+        cursor: pointer;
+    }
+
+    .hiddenMenu ul li{
+        vertical-align: center;
+        transition: 0s;
+        text-decoration: underline rgba(255,255,255,0);
+        text-underline-offset: 6px;
+    }
+
+    .menuLateral:focus-within .hiddenMenu{
+        z-index: 300;
+        height: auto;
+        opacity: 1;
+        padding: calc(4*var(--margem));
+    }
+
+/*Fim dos estilos do header*/
