+++ site/src/index.css
@@ -3,6 +3,7 @@ body {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
   font-family: Montserrat !important;
+  overflow-x: hidden;
 }
 
 code {
